overrides refactored as concerns, overrides_spec strings added
