+++ config/initializers/decidim_overrides.rb
@@ -4,4 +4,6 @@ Rails.application.config.to_prepare do
   Decidim::Initiatives::InitiativeMCell.prepend Decidim::Overrides::Initiatives::InitiativeMCell
   Decidim::Accountability::Result.include(Decidim::Accountability::ResultOverride)
   Decidim::Accountability::ResultsCalculator.include(Decidim::Accountability::ResultsCalculatorOverride)
+  Decidim::Meetings::Meeting.include(Decidim::Meetings::MeetingOverride)
+  Decidim::Meetings::MeetingsController.include(Decidim::Meetings::MeetingsControllerOverride)
 end
